Delete kesz

+++ model/entity/UserDto.cpp
@@ -10,6 +10,7 @@ UserDto::UserDto(QObject *parent) : QObject(parent)
 {
 
 }
+
 QString UserDto::username() const
 {
     return m_username;
@@ -20,6 +21,7 @@ void UserDto::setUsername(const QString &username)
     m_username = username;
     emit usernameChanged();
 }
+
 QString UserDto::password() const
 {
     return m_Password;
@@ -30,16 +32,18 @@ void UserDto::setPassword(const QString &password)
     m_Password = password;
     emit passwordChanged();
 }
-quint64 UserDto::id() const
+
+qint32 UserDto::id() const
 {
     return m_id;
 }
 
-void UserDto::setId(const quint64 &id)
+void UserDto::setId(const qint32 &id)
 {
     m_id = id;
     emit idChanged();
 }
+
 bool UserDto::active() const
 {
     return m_active;
